pyproject.toml
[tool.poetry]
name = "phoenix"
version = "0.1.0"
description = ""
authors = ["Alasdair <alnessy@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pyrsistent = "^0.16.0"
multipledispatch = "^0.6.0"
attrs = "^19.3.0"
janus = "^0.5.0"
cattrs = "^1.0.0"
sqlalchemy = "^1.3.17"
sqlalchemy-aio = "^0.15.0"
yarl = "^1.4.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
